requirements.txt
Flask
python-docx
PyPDF2
beautifulsoup4
requests
azure-ai-textanalytics
docx
python-dotenv
openai